backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
asyncpg
pydantic-settings
email-validator
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
